add validation for views

+++ products/views.py
@@ -48,6 +48,10 @@ def view_product_details(request, product_id):
 
 
 def create_product(request):
+    # validation of username
+    if not request.user.username == ('admin'):
+        return redirect(reverse(show_products))
+
     if request.method == 'POST':
         create_product_form = ProductForm(request.POST)
 
@@ -70,6 +74,10 @@ def create_product(request):
 
 
 def update_product(request, product_id):
+    # validation of username
+    if not request.user.username == ('admin'):
+        return redirect(reverse(show_products))
+
     # Get product to be updated from db
     product_being_updated = get_object_or_404(Product, pk=product_id)
 
@@ -100,6 +108,10 @@ def update_product(request, product_id):
 
 
 def delete_product(request, product_id):
+    # validation of username
+    if not request.user.username == ('admin'):
+        return redirect(reverse(show_products))
+
     product_to_delete = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         product_to_delete.delete()
